// ui/angular/src/app/core/api/ApiClient.ts
export class UpdateRoleRequest implements IUpdateRoleRequest {
    name!: string;
    description?: string | undefined;
    isDefault?: boolean;
    permissions?: string[] | undefined;

    constructor(data?: IUpdateRoleRequest) {
        if (data) {
            for (var property in data) {
                if (data.hasOwnProperty(property))
                    (<any>this)[property] = (<any>data)[property];
            }
        }
    }

    init(_data?: any) {
        if (_data) {
            this.name = _data["name"];
            this.description = _data["description"];
            this.isDefault = _data["isDefault"];
            if (Array.isArray(_data["permissions"])) {
                this.permissions = [] as any;
                for (let item of _data["permissions"])
                    this.permissions!.push(item);
            }
        }
    }

    static fromJS(data: any): UpdateRoleRequest {
        data = typeof data === 'object' ? data : {};
        let result = new UpdateRoleRequest();
        result.init(data);
        return result;
    }

    toJSON(data?: any) {
        data = typeof data === 'object' ? data : {};
        data["name"] = this.name;
        data["description"] = this.description;
        data["isDefault"] = this.isDefault;
        if (Array.isArray(this.permissions)) {
            data["permissions"] = [];
            for (let item of this.permissions)
                data["permissions"].push(item);
        }
        return data;
    }
}

export interface IUpdateRoleRequest {
    name: string;
    description?: string | undefined;
    isDefault?: boolean;
    permissions?: string[] | undefined;
}
// ...
export class AddRoleRequest implements IAddRoleRequest {
    name!: string;
    description?: string | undefined;
    isDefault?: boolean;
    permissions?: string[] | undefined;

    constructor(data?: IAddRoleRequest) {
        if (data) {
            for (var property in data) {
                if (data.hasOwnProperty(property))
                    (<any>this)[property] = (<any>data)[property];
            }
        }
    }

    init(_data?: any) {
        if (_data) {
            this.name = _data["name"];
            this.description = _data["description"];
            this.isDefault = _data["isDefault"];
            if (Array.isArray(_data["permissions"])) {
                this.permissions = [] as any;
                for (let item of _data["permissions"])
                    this.permissions!.push(item);
            }
        }
    }

    static fromJS(data: any): AddRoleRequest {
        data = typeof data === 'object' ? data : {};
        let result = new AddRoleRequest();
        result.init(data);
        return result;
    }

    toJSON(data?: any) {
        data = typeof data === 'object' ? data : {};
        data["name"] = this.name;
        data["description"] = this.description;
        data["isDefault"] = this.isDefault;
        if (Array.isArray(this.permissions)) {
            data["permissions"] = [];
            for (let item of this.permissions)
                data["permissions"].push(item);
        }
        return data;
    }
}

export interface IAddRoleRequest {
    name: string;
    description?: string | undefined;
    isDefault?: boolean;
    permissions?: string[] | undefined;
}

export class PermissionGroup implements IPermissionGroup {
    groupId?: string | undefined;
    groupName?: string | undefined;
    items?: Permission[] | undefined;

    constructor(data?: IPermissionGroup) {
        if (data) {
            for (var property in data) {
                if (data.hasOwnProperty(property))
                    (<any>this)[property] = (<any>data)[property];
            }
        }
    }

    init(_data?: any) {
        if (_data) {
            this.groupId = _data["groupId"];
            this.groupName = _data["groupName"];
            if (Array.isArray(_data["items"])) {
                this.items = [] as any;
                for (let item of _data["items"])
                    this.items!.push(Permission.fromJS(item));
            }
        }
    }

    static fromJS(data: any): PermissionGroup {
        data = typeof data === 'object' ? data : {};
        let result = new PermissionGroup();
        result.init(data);
        return result;
    }

    toJSON(data?: any) {
        data = typeof data === 'object' ? data : {};
        data["groupId"] = this.groupId;
        data["groupName"] = this.groupName;
        if (Array.isArray(this.items)) {
            data["items"] = [];
            for (let item of this.items)
                data["items"].push(item.toJSON());
        }
        return data;
    }
}

export interface IPermissionGroup {
    groupId?: string | undefined;
    groupName?: string | undefined;
    items?: Permission[] | undefined;
}

export class Permission implements IPermission {
    id?: string | undefined;
    name?: string | undefined;
    visibleName?: string | undefined;

    constructor(data?: IPermission) {
        if (data) {
            for (var property in data) {
                if (data.hasOwnProperty(property))
                    (<any>this)[property] = (<any>data)[property];
            }
        }
    }

    init(_data?: any) {
        if (_data) {
            this.id = _data["id"];
            this.name = _data["name"];
            this.visibleName = _data["visibleName"];
        }
    }

    static fromJS(data: any): Permission {
        data = typeof data === 'object' ? data : {};
        let result = new Permission();
        result.init(data);
        return result;
    }

    toJSON(data?: any) {
        data = typeof data === 'object' ? data : {};
        data["id"] = this.id;
        data["name"] = this.name;
        data["visibleName"] = this.visibleName;
        return data;
    }
}

export interface IPermission {
    id?: string | undefined;
    name?: string | undefined;
    visibleName?: string | undefined;
}